# Consolidate passes through path ast